Moved request to a better place
