+++ src/components/file-previews/AudioPreview/AudioPreview.jsx
@@ -1,21 +1,13 @@
-import React, {
-  useImperativeHandle,
-  useRef,
-  forwardRef,
-  useState,
-  useEffect
-} from 'react';
+import React, { useImperativeHandle, useRef, forwardRef } from 'react';
 import DefaultFileTitle from '../components/DefaultFileTitle/DefaultFileTitle';
 import DefaultFileActions from '../components/DefaultFileActions/DefaultFileActions';
 import AudioPlayer from '../components/AudioPlayer/AudioPlayer';
-import { getFilePreviewEffect } from '../../../effects/filesEffects';
 import styles from './AudioPreview.module.scss';
 
 const AudioPreview = forwardRef(
   (
     {
       mode = 'default',
-      usePreviewImage = true,
       file,
       fileContent,
       fileContentType,
@@ -25,7 +17,6 @@ const AudioPreview = forwardRef(
     ref
   ) => {
     const audioRef = useRef(null);
-    const [filePreviewImage, setFilePreviewImage] = useState(null);
 
     useImperativeHandle(ref, () => ({
       runPreview: () => {
@@ -40,25 +31,6 @@ const AudioPreview = forwardRef(
       }
     }));
 
-    useEffect(() => {
-      if (usePreviewImage) {
-        fetchPreview();
-      }
-    }, []);
-
-    const fetchPreview = async () => {
-      try {
-        const preview = await getFilePreviewEffect(
-          file.slug,
-          null,
-          file.extension
-        );
-        setFilePreviewImage(preview);
-      } catch (e) {
-        setFilePreviewImage(null);
-      }
-    };
-
     return (
       <div className={styles.container}>
         <AudioPlayer
@@ -66,7 +38,6 @@ const AudioPreview = forwardRef(
           file={file}
           fileContent={fileContent}
           fileContentType={fileContentType}
-          filePreviewImage={filePreviewImage}
         />
 
         {mode === 'default' && (
